+++ src/test/groovy/ru/vyarus/gradle/plugin/github/configurer/BintrayConfigurerTest.groovy
@@ -12,6 +12,7 @@ class BintrayConfigurerTest extends AbstractTest {
     def "Check configuration applied"() {
 
         when: "plugin configured"
+        file('CHANGELOG.md').createNewFile()
         Project project = project {
             apply plugin: 'java'
             apply plugin: 'ru.vyarus.github-info'
@@ -29,11 +30,14 @@ class BintrayConfigurerTest extends AbstractTest {
         project.bintray.pkg.vcsUrl == "https://github.com/${projectId}.git"
         project.bintray.pkg.issueTrackerUrl == "https://github.com/${projectId}/issues"
         project.bintray.pkg.licenses == ['MIT']
+        project.bintray.pkg.githubRepo == projectId
+        project.bintray.pkg.githubReleaseNotesFile == 'CHANGELOG.md'
     }
 
     def "Check user configuration preserved"() {
 
         when: "plugin configured"
+        file('CHANGELOG.md').createNewFile()
         Project project = project {
             apply plugin: 'java'
             apply plugin: 'ru.vyarus.github-info'
@@ -48,6 +52,7 @@ class BintrayConfigurerTest extends AbstractTest {
                 pkg {
                     websiteUrl = 'http://google.com'
                     licenses = ['OTHER']
+                    githubReleaseNotesFile = 'FOO.txt'
                 }
             }
         }
@@ -58,5 +63,7 @@ class BintrayConfigurerTest extends AbstractTest {
         project.bintray.pkg.vcsUrl == "https://github.com/${projectId}.git"
         project.bintray.pkg.issueTrackerUrl == "https://github.com/${projectId}/issues"
         project.bintray.pkg.licenses == ['OTHER']
+        project.bintray.pkg.githubRepo == projectId
+        project.bintray.pkg.githubReleaseNotesFile == 'FOO.txt'
     }
 }