+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+* Add `changelogFile` configuration property, used to specify changelog file path. Default file name is recognized by 
+    CHANGELOG.md, CHANGELOG.txt or CHANGELOG files in project root.
+* Add bintray plugin properties support (available since 1.6 plugin version):
+    - githubRepo (required for bintray github integration and to show github readme)
+    - githubReleaseNotesFile (used to show github changelog on bintray)
+
 ### 1.0.0 (2015-12-05)
 * Initial release